requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
motor==3.1.1
pydantic==1.10.11
passlib[bcrypt]==1.7.4
python-jose==3.3.0
email-validator==2.0.0.post3
httpx==0.24.0
python-multipart==0.0.6
